backend/Recommend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
Django==4.0.2
djangorestframework==3.13.1
elasticsearch==7.17.0
mysqlclient==2.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
